requirements.txt
django=4.1.11
djangorestframework=3.13.1
